refactor(runtime): separate source query and reactive execution
