+++ src/runtime/src/runtime/execution/query.rs
@@ -1,0 +1,129 @@
+use super::*;
+
+impl MechRuntime {
+  #[cfg(feature = "invariant_define")]
+  pub fn integrity_constraint_report(
+    &self,
+  ) -> MResult<mech_program::IntegrityConstraintReport> {
+    self.program.integrity_constraint_report()
+  }
+
+  pub fn out_string(&self) -> String {
+    self.program.out_string()
+  }
+
+  pub fn has_interpreter(&self, interpreter_id: u64) -> bool {
+    self.program.has_interpreter(interpreter_id)
+  }
+
+  pub fn root_plan_len(&self) -> usize {
+    self.program.interpreter().plan_len()
+  }
+
+  pub fn output_value_for_interpreter(
+    &self,
+    interpreter_id: u64,
+    output_id: u64,
+  ) -> Option<RuntimeValueSnapshot> {
+    self
+      .program
+      .output_value_for_interpreter(interpreter_id, output_id)
+      .as_ref()
+      .map(RuntimeValueSnapshot::capture)
+  }
+
+  pub fn symbol_name_for_interpreter_output(
+    &self,
+    interpreter_id: u64,
+    output_id: u64,
+  ) -> Option<String> {
+    self.program.symbol_name_for_interpreter_output(interpreter_id, output_id)
+  }
+
+  pub fn symbol_values_for_interpreter(
+    &self,
+    interpreter_id: u64,
+    names: &[String],
+  ) -> Option<Vec<(String, RuntimeValueSnapshot)>> {
+    self
+      .program
+      .symbol_values_for_interpreter(interpreter_id, names)
+      .map(|values| {
+        values
+          .into_iter()
+          .map(|(name, value)| {
+            (name, RuntimeValueSnapshot::capture(&value))
+          })
+          .collect()
+      })
+  }
+
+  pub fn root_symbol_value(
+    &self,
+    name: &str,
+  ) -> MResult<RuntimeValueSnapshot> {
+    self
+      .program
+      .root_symbol_value(name)
+      .map(|value| RuntimeValueSnapshot::capture(&value))
+  }
+
+  pub fn root_symbol_values(
+    &self,
+    names: &[&str],
+  ) -> MResult<Vec<(String, RuntimeValueSnapshot)>> {
+    self
+      .program
+      .root_symbol_values(names)
+      .map(|values| {
+        values
+          .into_iter()
+          .map(|(name, value)| {
+            (name, RuntimeValueSnapshot::capture(&value))
+          })
+          .collect()
+      })
+  }
+
+  pub fn root_symbol_values_all(
+    &self,
+  ) -> Vec<(String, RuntimeValueSnapshot)> {
+    self
+      .program
+      .root_symbol_values_all()
+      .into_iter()
+      .map(|(name, value)| {
+        (name, RuntimeValueSnapshot::capture(&value))
+      })
+      .collect()
+  }
+
+  pub fn bind_ans_for_interpreter(
+    &mut self,
+    interpreter_id: u64,
+    value: &Value,
+  ) -> MResult<()> {
+    if let Some(owner) = self.program_transaction_owner {
+      return Err(MechError::new(
+        RuntimeProgramBusy {
+          operation: "bind_ans_for_interpreter",
+          owner,
+          requester: None,
+        },
+        None,
+      ));
+    }
+
+    if self.program.bind_ans_for_interpreter(interpreter_id, value) {
+      return Ok(());
+    }
+
+    Err(MechError::new(
+      RuntimeInvalidOperationError {
+        operation: "bind_ans_for_interpreter",
+        reason: format!("interpreter id {} not found", interpreter_id),
+      },
+      None,
+    ))
+  }
+}
